Update README.md
Preceding law will eventually prevail as it relates to intellectual property (IP) - the discoverer who can PROVE he was first, is first. There are some fairly recent IP laws and/or statutes which are repugnant to precedence and therefore must and will have been struck down per lineage of law. Lineage of law, being itself ancient, was further cemented by the supreme Law of the Land AKA the Supremacy Clause of the US Constitution. When this fact is publicly recognized in the judiciary ALL APPLICABLE CONTENTS OF THIS FORUM will continue to prevail as written and public testimony to the fact that I, Josef Kulovany, was first as it relates to the entirety of IP posted by myself herein. This is the proper course of things per the patent office mission statement - “The mission of the U.S. Patent and Trademark Office (USPTO) is to foster innovation, competitiveness and economic growth, domestically and abroad, by providing high quality and timely examination of patent and trademark applications, guiding domestic and international intellectual property (IP) policy, and delivering IP information and education worldwide.”

The supreme Law of the Land AKA The Supremacy Clause cemented the US Constitution as law, and had long since re-established lineage of law as it relates to the common law organization of precedence in my country of residence, the United States, which is formally and internationally recognized by a myriad of treaties and otherwise to be a sovereign entity whose laws are recognized therein by every established foreign nation on this earth. Tread carefully, all ye vipers and snakes.

I hereby retroactively, proactively, and presently do not grant release in any way, shape, or form of any and all applicable contents of this public forum, now, in the future, nor in the past without express written permission of myself, Josef Kulovany.

Notwithstanding, all applicable copyright contained herein past, present, future also stands. ALL RIGHTS RESERVED. You MUST seek the express written permission of Josef Kulovany in order to proceed with copying any and all portions of the contents of this forum past, present, and future per applicable laws, treaties, and otherwise.

YOU HAVE BEEN WARNED not only by this notice, but pursuant to preceding and also supreme Law, which is binding according to a myriad of treaties and otherwise - my own failure to enumerate, of course, in no way, shape, or form revokes the legitimacy of these claims, and all rights held by myself, Josef Kulovany, which are not claimed herein. As before, ALL RIGHTS RESERVED.

WRITE TO ME FOR APPLICABLE PERMISSIONS - charg.chg.wecharg@gmail.com.